first steps towards real MPNN based solubility model
